Check OpenAPI attributes. Split chat history delete endpoint to work around OpenAPI not supporting optional path params.

+++ src/Api/ChatHistory_v1.cs
@@ -30,8 +30,16 @@ namespace Aire.Memory.Api
         }
 
         [Function("GetChatHistory_v1")]
-        [OpenApiOperation(operationId: "GetChatHistory", tags: ["chat-history"], Description = "Get a list of chat logs")]
-        [OpenApiSecurity("bearer_auth", SecuritySchemeType.Http, Scheme = OpenApiSecuritySchemeType.Bearer, BearerFormat = "JWT", Description = "User token")]
+        [OpenApiOperation(
+            operationId: "getChatHistory", 
+            tags: ["chat-history"], 
+            Summary = "Get a list of chat logs")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth", 
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
         [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(List<ChatLogMetadata>), Description = "List of chat metadata objects")]
         [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
         public async Task<IActionResult> GetChatHistory(
@@ -51,9 +59,17 @@ namespace Aire.Memory.Api
         }
 
         [Function("GetChatHistoryWithId_v1")]
-        [OpenApiOperation(operationId: "GetChatHistoryWithId", tags: ["chat-history"], Description = "Retrieve a chat log")]
-        [OpenApiSecurity("bearer_auth", SecuritySchemeType.Http, Scheme = OpenApiSecuritySchemeType.Bearer, BearerFormat = "JWT", Description = "User token")]
-        [OpenApiParameter("id", Description = "Chat log identifier", Required = true)]
+        [OpenApiOperation(
+            operationId: "getChatHistoryWithId", 
+            tags: ["chat-history"], 
+            Summary = "Retrieve a chat log")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth", 
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
+        [OpenApiParameter("id", Description = "Chat log identifier", In = ParameterLocation.Path, Required = true)]
         [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(List<ChatMessage>), Description = "List of chat messages")]
         [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found.")]
         [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid parameter")]
@@ -82,8 +98,16 @@ namespace Aire.Memory.Api
         }
 
         [Function("PostChatHistory_v1")]
-        [OpenApiOperation(operationId: "PostChatHistory", tags: ["chat-history"], Description = "Store new chat log")]
-        [OpenApiSecurity("bearer_auth", SecuritySchemeType.Http, Scheme = OpenApiSecuritySchemeType.Bearer, BearerFormat = "JWT", Description = "User token")]
+        [OpenApiOperation(
+            operationId: "postChatHistory", 
+            tags: ["chat-history"], 
+            Summary = "Store new chat log")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth", 
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
         [OpenApiRequestBody("application/json", typeof(List<ChatMessage>), Description = "List of chat messages", Required = true)]
         [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(ChatLogMetadata), Description = "Chat log metadata")]
         [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid body")]
@@ -115,8 +139,16 @@ namespace Aire.Memory.Api
         }
 
         [Function("PutChatHistory_v1")]
-        [OpenApiOperation(operationId: "PutChatHistory", tags: ["chat-history"], Description = "Edit existing chat log")]
-        [OpenApiSecurity("bearer_auth", SecuritySchemeType.Http, Scheme = OpenApiSecuritySchemeType.Bearer, BearerFormat = "JWT", Description = "User token")]
+        [OpenApiOperation(
+            operationId: "putChatHistory", 
+            tags: ["chat-history"], 
+            Summary = "Edit existing chat log")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth", 
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
         [OpenApiParameter("id", Description = "Chat log identifier", Required = true)]
         [OpenApiRequestBody("application/json", typeof(List<ChatMessage>), Description = "List of chat messages", Required = true)]
         [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(ChatLogMetadata), Description = "Chat log metadata")]
@@ -156,17 +188,61 @@ namespace Aire.Memory.Api
         }
 
         [Function("DeleteChatHistory_v1")]
-        [OpenApiOperation(operationId: "DeleteChatHistory", tags: ["chat-history"], Description = "Wipe a chat log")]
-        [OpenApiSecurity("bearer_auth", SecuritySchemeType.Http, Scheme = OpenApiSecuritySchemeType.Bearer, BearerFormat = "JWT", Description = "User token")]
-        [OpenApiParameter("id", Description = "Chat log identifier", Required = true)]
+        [OpenApiOperation(
+            operationId: "deleteChatHistory", 
+            tags: ["chat-history"], 
+            Summary = "Delete entire chat history")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth", 
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
         [OpenApiResponseWithoutBody(HttpStatusCode.NoContent, Description = "The chatlog(s) removed successfully")]
         [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found")]
         [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param")]
         [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
         public async Task<IActionResult> DeleteChatHistory(
-            [HttpTrigger(AuthorizationLevel.Function, "delete", Route = "v1/chat-history/{id?}")] HttpRequest req,
+            [HttpTrigger(AuthorizationLevel.Function, "delete", Route = "v1/chat-history")] HttpRequest req,
+            FunctionContext context)
+        {
+            var auth = context.Features.Get<JwtAuthFeature>();
+            if(!_jwt.CheckAuthorization(auth, AireRoles.User, AireScopes.DeleteChatHistory))
+                return new UnauthorizedResult();
+
+            var history = await _db.ChatLogs
+                .Where(x => x.UserId == auth!.User)
+                .ToListAsync();
+
+            foreach(var chat in history)
+            {
+                _db.ChatLogs.Remove(chat);
+            }
+            await _db.SaveChangesAsync();
+
+            return new NoContentResult();
+        }
+
+        [Function("DeleteChatHistoryWithId_v1")]
+        [OpenApiOperation(
+            operationId: "deleteChatHistoryWithId", 
+            tags: ["chat-history"], 
+            Summary = "Delete a chat log")]
+        [OpenApiSecurity(
+            schemeName: "bearer_auth",
+            schemeType: SecuritySchemeType.Http, 
+            Scheme = OpenApiSecuritySchemeType.Bearer, 
+            BearerFormat = "JWT", 
+            Description = "User token")]
+        [OpenApiParameter("id", Description = "Chat log identifier", In = ParameterLocation.Path, Required = true)]
+        [OpenApiResponseWithoutBody(HttpStatusCode.NoContent, Description = "The chatlog(s) removed successfully")]
+        [OpenApiResponseWithoutBody(HttpStatusCode.NotFound, Description = "The chat log was not found")]
+        [OpenApiResponseWithoutBody(HttpStatusCode.BadRequest, Description = "Invalid param")]
+        [OpenApiResponseWithoutBody(HttpStatusCode.Unauthorized, Description = "Missing or insufficient authorization")]
+        public async Task<IActionResult> DeleteChatHistoryWithId(
+            [HttpTrigger(AuthorizationLevel.Function, "delete", Route = "v1/chat-history/{id}")] HttpRequest req,
             FunctionContext context,
-            string? id = null)
+            string id)
         {
             var auth = context.Features.Get<JwtAuthFeature>();
             if(!_jwt.CheckAuthorization(auth, AireRoles.User, AireScopes.DeleteChatHistory))
@@ -175,28 +251,14 @@ namespace Aire.Memory.Api
             if(!Guid.TryParse(id, out Guid chatId))
                 return new BadRequestResult();
 
-            List<ChatLogEntity> toBeRemoved;
+            var chat = await _db.ChatLogs
+                .Where(x => x.UserId == auth!.User && x.Id == chatId)
+                .FirstOrDefaultAsync();
 
-            if(id == null)
-            {
-                toBeRemoved = await _db.ChatLogs
-                    .Where(x => x.UserId == auth!.User)
-                    .ToListAsync();
-            }
-            else
-            {
-                toBeRemoved = await _db.ChatLogs
-                    .Where(x => x.Id == chatId && x.UserId == auth!.User)
-                    .ToListAsync();
+            if(chat == null)
+                return new NotFoundResult();
 
-                if(toBeRemoved.Count == 0)
-                    return new NotFoundResult();
-            }
-
-            foreach(var chat in toBeRemoved)
-            {
-                _db.ChatLogs.Remove(chat);
-            }
+            _db.ChatLogs.Remove(chat);
             await _db.SaveChangesAsync();
 
             return new NoContentResult();
